Rename module `clocks` -> `clocky`.
