+++ src/async/export/TokenBucket.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { IntfTimeSource, StdTimeSource } from '@this/clocks';
+import { IntfTimeSource, StdTimeSource } from '@this/clocky';
 import { Duration, Frequency, Moment } from '@this/data-values';
 import { MustBe } from '@this/typey';
