+++ src/async/tests/TokenBucket.test.js
@@ -5,7 +5,7 @@ import { setImmediate } from 'node:timers/promises';
 
 import { PromiseState, TokenBucket } from '@this/async';
 import { IntfTimeSource, MockTimeSource, StdTimeSource }
-  from '@this/clocks';
+  from '@this/clocky';
 import { Duration, Frequency, Moment } from '@this/data-values';
