+++ src/clocky/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/clocks",
+  "name": "@this/clocky",
   "version": "0.0.1",
   "type": "module",
   "private": true,
