+++ src/host/private/ShutdownHandler.js
@@ -4,7 +4,7 @@
 import process from 'node:process'; // Need to import as such, for `.on*()`.
 
 import { Threadlet } from '@this/async';
-import { WallClock } from '@this/clocks';
+import { WallClock } from '@this/clocky';
 import { IntfLogger } from '@this/loggy-intf';
 
 import { CallbackList } from '#p/CallbackList';
