+++ src/main-lactoserv/package.json
@@ -13,7 +13,7 @@
 
   "dependencies": {
     "@this/async": "*",
-    "@this/clocks": "*",
+    "@this/clocky": "*",
     "@this/data-values": "*",
     "@this/host": "*",
     "@this/loggy": "*",
