+++ src/main-lactoserv/private/Debugging.js
@@ -4,7 +4,7 @@
 import { setImmediate as setImmediateCallback } from 'node:timers';
 import { setImmediate } from 'node:timers/promises';
 
-import { WallClock } from '@this/clocks';
+import { WallClock } from '@this/clocky';
 import { Duration } from '@this/data-values';
 import { Host } from '@this/host';
 import { IntfLogger } from '@this/loggy-intf';
