+++ src/webapp-builtins/package.json
@@ -14,7 +14,7 @@
 
   "dependencies": {
     "@this/async": "*",
-    "@this/clocks": "*",
+    "@this/clocky": "*",
     "@this/collections": "*",
     "@this/compy": "*",
     "@this/data-values": "*",
